refactor: specify model type in model file names
