+++ src/app/app-routing.module.ts
@@ -3,7 +3,7 @@ import { RouterModule, Routes } from '@angular/router';
 
 import { AuthenticationGuard } from './core/guards/authentication/authentication.guard';
 import { AuthorizationGuard } from './core/guards/authorization/authorization.guard';
-import { UserRole } from './core/models/user/user-role.model';
+import { UserRole } from './core/models/user/user-role.enum';
 
 const routes: Routes = [
   {
