+++ src/app/core/guards/authorization/authorization.guard.ts
@@ -3,7 +3,7 @@ import { CanActivateFn } from '@angular/router';
 import { Store } from '@ngrx/store';
 import { Observable, map } from 'rxjs';
 
-import { UserRole } from 'app/core/models/user/user-role.model';
+import { UserRole } from 'app/core/models/user/user-role.enum';
 import { selectAuthState } from 'app/state/auth/auth.selectors';
 import { AuthSessionState } from 'app/state/auth/auth.feature';
