+++ src/app/core/models/user/user.interface.ts
@@ -1,4 +1,4 @@
-import { UserRole } from './user-role.model';
+import { UserRole } from './user-role.enum';
 
 export interface User {
   id: string;
