+++ src/app/core/services/auth/auth.service.spec.ts
@@ -8,10 +8,10 @@ import { HttpClient } from '@angular/common/http';
 import { AuthService } from './auth.service';
 import { LocalStorageService } from '../local-storage/local-storage.service';
 import { AuthSession } from 'app/state/auth/auth.feature';
-import { UserRole } from 'app/core/models/user/user-role.model';
-import { SignupCredentials } from 'app/core/models/auth/signup.model';
-import { User } from 'app/core/models/user/user.model';
-import { LoginCredentials } from 'app/core/models/auth/login.model';
+import { UserRole } from 'app/core/models/user/user-role.enum';
+import { SignupCredentials } from 'app/core/models/auth/signup.interface';
+import { User } from 'app/core/models/user/user.interface';
+import { LoginCredentials } from 'app/core/models/auth/login.interface';
 
 describe('AuthService', () => {
   let authService: AuthService;
