+++ src/app/pages/auth/components/signup-form/signup-form.component.ts
@@ -4,7 +4,7 @@ import { Store } from '@ngrx/store';
 
 import { FormBaseComponent } from 'app/shared/components/form-base/form-base.component';
 import { SelectOption } from 'app/shared/components/select-field/select-field.component';
-import { UserRole } from 'app/core/models/user/user-role.model';
+import { UserRole } from 'app/core/models/user/user-role.enum';
 import { SentenceCasePipe } from 'app/shared/pipes/sentence-case/sentence-case.pipe';
 import { SplitCamelCasePipe } from 'app/shared/pipes/split-camel-case/split-camel-case.pipe';
 import { EmailValidator } from 'app/core/validators/email/email.validator';
