+++ src/app/shared/components/user-button/user-button.component.ts
@@ -2,7 +2,7 @@ import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { Observable, map } from 'rxjs';
 
-import { User } from 'app/core/models/user/user.model';
+import { User } from 'app/core/models/user/user.interface';
 import { selectLoggedUser } from 'app/state/auth/auth.selectors';
 import { AuthActions } from 'app/state/auth/auth.actions';
