+++ src/app/state/auth/auth.feature.ts
@@ -1,6 +1,6 @@
 import { createFeature, createReducer, on } from '@ngrx/store';
 
-import { User } from 'app/core/models/user/user.model';
+import { User } from 'app/core/models/user/user.interface';
 import { AuthActions } from './auth.actions';
 import { AuthActionReducers } from './auth.reducers';
