+++ src/app/state/auth/auth.reducers.spec.ts
@@ -1,6 +1,6 @@
-import { User } from 'app/core/models/user/user.model';
+import { User } from 'app/core/models/user/user.interface';
 import { AuthActionReducers } from './auth.reducers';
-import { UserRole } from 'app/core/models/user/user-role.model';
+import { UserRole } from 'app/core/models/user/user-role.enum';
 import { AuthSession } from './auth.feature';
 import { AuthActions } from './auth.actions';
